model deck struct & deck functions
Put array of 108 allocated cards in a deck struct. This sort of
formalizes the idea that a deck has 108 cards. Anything that receives a
deck knows it has space for no more than 108 cards. This is better than
functions receiving just an array of cards and presuming it has a length
of 108.